# Address build_runner performance issues