--- a/lib/src/app.dart
+++ b/lib/src/app.dart
@@ -7,6 +7,7 @@ import 'package:lichess_mobile/l10n/l10n.dart';
 import 'package:lichess_mobile/src/constants.dart';
 import 'package:lichess_mobile/src/model/account/account_repository.dart';
 import 'package:lichess_mobile/src/model/challenge/challenge_service.dart';
+import 'package:lichess_mobile/src/model/common/preloaded_data.dart';
 import 'package:lichess_mobile/src/model/correspondence/correspondence_service.dart';
 import 'package:lichess_mobile/src/model/notifications/notification_service.dart';
 import 'package:lichess_mobile/src/model/settings/board_preferences.dart';
@@ -18,6 +19,35 @@ import 'package:lichess_mobile/src/network/http.dart';
 import 'package:lichess_mobile/src/network/socket.dart';
 import 'package:lichess_mobile/src/styles/styles.dart';
 import 'package:lichess_mobile/src/utils/screen.dart';
+
+/// Application initialization and main entry point.
+class AppInitializationScreen extends ConsumerWidget {
+  const AppInitializationScreen({super.key});
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    ref.listen<AsyncValue<PreloadedData>>(
+      preloadedDataProvider,
+      (_, state) {
+        if (state.hasValue || state.hasError) {
+          FlutterNativeSplash.remove();
+        }
+      },
+    );
+
+    return ref.watch(preloadedDataProvider).when(
+          data: (_) => const Application(),
+          // loading screen is handled by the native splash screen
+          loading: () => const SizedBox.shrink(),
+          error: (err, st) {
+            debugPrint(
+              'SEVERE: [App] could not initialize app; $err\n$st',
+            );
+            return const SizedBox.shrink();
+          },
+        );
+  }
+}
 
 /// The main application widget.
 ///
@@ -38,8 +68,6 @@ class _AppState extends ConsumerState<Application> {
 
   @override
   void initState() {
-    FlutterNativeSplash.remove();
-
     _appLifecycleListener = AppLifecycleListener(
       onResume: () async {
         final online = await isOnline(ref.read(defaultClientProvider));