--- a/lib/src/init.dart
+++ b/lib/src/init.dart
@@ -7,7 +7,7 @@ import 'package:flutter_displaymode/flutter_displaymode.dart';
 import 'package:lichess_mobile/src/binding.dart';
 import 'package:lichess_mobile/src/db/secure_storage.dart';
 import 'package:lichess_mobile/src/model/settings/board_preferences.dart';
-import 'package:lichess_mobile/src/model/settings/preferences.dart';
+import 'package:lichess_mobile/src/model/settings/preferences_storage.dart';
 import 'package:lichess_mobile/src/network/socket.dart';
 import 'package:lichess_mobile/src/utils/chessboard.dart';
 import 'package:lichess_mobile/src/utils/color_palette.dart';