--- a/lib/src/model/analysis/analysis_controller.dart
+++ b/lib/src/model/analysis/analysis_controller.dart
@@ -749,7 +749,7 @@ class AnalysisState with _$AnalysisState {
   bool get canGoNext => currentNode.hasChild;
   bool get canGoBack => currentPath.size > UciPath.empty.size;
 
-  EngineGaugeParams get engineGaugeParams => EngineGaugeParams(
+  EngineGaugeParams get engineGaugeParams => (
         orientation: pov,
         isLocalEngineAvailable: isEngineAvailable,
         position: position,