BEFORE:
`dart run build_runner build`
```
[INFO] Running build completed, took 41.3s
[INFO] Caching finalized dependency graph completed, took 252ms
[INFO] Succeeded after 41.6s with 477 outputs (2326 actions)
```

`dart run build_runner watch` with a change in `src/view/settings/toggle_sound_button.dart`:
```
[INFO] Running build completed, took 8.7s
[INFO] Caching finalized dependency graph completed, took 207ms
[INFO] Succeeded after 8.9s with 205 outputs (473 actions)
```


AFTER:

`dart run build_runner build`
```
[INFO] Running build completed, took 26.5s
[INFO] Caching finalized dependency graph completed, took 247ms
[INFO] Succeeded after 26.7s with 443 outputs (1407 actions)
```

`dart run build_runner watch` with a change in `src/view/settings/toggle_sound_button.dart`:
```
[INFO] Running build completed, took 1.1s
[INFO] Caching finalized dependency graph completed, took 95ms
[INFO] Succeeded after 1.2s with 18 outputs (30 actions)
```